Make ToolboxEditorHandler public; minor refactor changes
