+++ Assets/Editor Toolbox/Editor/ToolboxEditorDrawer.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-
 using UnityEditor;
 
 namespace Toolbox.Editor
